Level compiler: add changes (from SporkHack, rev. 588, 590-592)
- ROOM and SUBROOM room type chance shouldn't generate opcodes for if-statement
- Fix obscure 64-bit bug that caused crash while creating Sanctum: use long, not int
- Cast ints etc to long for add_opvars(), in case sizeof int != sizeof long
- Get rid of current_region/current_coord hacks, and use yacc features instead

+++ util/lev_main.c
@@ -687,8 +687,8 @@ sp_lev *sp;
 {
 	register int i, len;
 	register char *s1, *s2;
-	int max_len = 0;
-	int max_hig = 0;
+	long max_len = 0;
+	long max_hig = 0;
 	char msg[256];
 	char *tmpmap[ROWNO];
 	int dx,dy;
